rename free c string; start wrapper util on python side
